refactor(frontend/login): create login module
