+++ packages/frontend/src/components/containers/index.ts
@@ -1,8 +1,7 @@
 import RegisterForm from './RegisterForm';
-import LoginForm from './LoginForm';
 import PayPal from './PayPal';
 import CartContent from './CartContent';
 import UserForms from './UserForms';
 import UserProducts from './UserProducts';
 
-export { RegisterForm, LoginForm, PayPal, CartContent, UserForms, UserProducts };
+export { RegisterForm, PayPal, CartContent, UserForms, UserProducts };
