+++ packages/frontend/src/components/modules/Login/LoginForm.tsx
@@ -1,15 +1,15 @@
 import React, { useState, SyntheticEvent } from 'react';
 import { Link, useHistory } from 'react-router-dom';
-import '../../styles/containers/registerForm.css';
-import '../../styles/toastify.css';
-import { post } from '../../utils/petitions';
+import '../../../styles/containers/registerForm.css';
+import '../../../styles/toastify.css';
+import { post } from '../../../utils/petitions';
 import Cookie from 'universal-cookie';
 import { ToastContainer, toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 
-import { Loader, PrimaryButton } from '../elements';
-import { Input } from '../elements';
-import { HTTPException } from '../../utils/HttpException';
+import { Loader, PrimaryButton } from '../../elements';
+import { Input } from '../../elements';
+import { HTTPException } from '../../../utils/HttpException';
 
 const LoginForm = () => {
   const [email, setEmail] = useState<string>('');
